Move convertUTCDateToLocalDate to utils
