+++ ui/src/util/convertUTCDateToLocalDate.js
@@ -1,0 +1,9 @@
+export default function convertUTCDateToLocalDate(date) {
+  const newDate = new Date(
+    date.getTime() + date.getTimezoneOffset() * 60 * 1000
+  );
+  const offset = date.getTimezoneOffset() / 60;
+  const hours = date.getHours();
+  newDate.setHours(hours - offset);
+  return newDate;
+}
